+++ bittensor/substrate/utils/ss58.py
@@ -1,6 +1,6 @@
 # Python Substrate Interface Library
 #
-# Copyright 2018-2021 Stichting Polkascan (Polkascan Foundation).
+# Copyright 2018-2020 Stichting Polkascan (Polkascan Foundation).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,96 +17,71 @@
 #  ss58.py
 
 """ SS58 is a simple address format designed for Substrate based chains.
-    Encoding/decoding according to specification on
-    https://github.com/paritytech/substrate/wiki/External-Address-Format-(SS58)
+    Encoding/decoding according to specification on https://wiki.parity.io/External-Address-Format-(SS58)
 
 """
-import warnings
-from typing import Optional
-
 import base58
 from hashlib import blake2b
 
-from scalecodec.base import ScaleBytes, ScaleDecoder
+from scalecodec import ScaleBytes
+from scalecodec.types import U8, U16, U32, U64
 
 
-def ss58_decode(address: str, valid_ss58_format: Optional[int] = None, valid_address_type=None) -> str:
+def ss58_decode(address, valid_address_type=None):
     """
     Decodes given SS58 encoded address to an account ID
     Parameters
     ----------
     address: e.g. EaG2CRhJWPb7qmdcJvy3LiWdh26Jreu9Dx6R1rXxPmYXoDk
-    valid_ss58_format
     valid_address_type
 
     Returns
     -------
     Decoded string AccountId
     """
-
-    # Check if address is already decoded
-    if address.startswith('0x'):
-        return address
-
-    if valid_address_type is not None:
-        warnings.warn("Keyword 'valid_address_type' will be replaced by 'valid_ss58_format'", DeprecationWarning)
-        valid_ss58_format = valid_address_type
-
     checksum_prefix = b'SS58PRE'
 
-    address_decoded = base58.b58decode(address)
+    ss58_format = base58.b58decode(address)
 
-    if address_decoded[0] & 0b0100_0000:
-        ss58_format_length = 2
-        ss58_format = ((address_decoded[0] & 0b0011_1111) << 2) | (address_decoded[1] >> 6) | \
-                      ((address_decoded[1] & 0b0011_1111) << 8)
-    else:
-        ss58_format_length = 1
-        ss58_format = address_decoded[0]
-
-    if ss58_format in [46, 47]:
-        raise ValueError(f"{ss58_format} is a reserved SS58 format")
-
-    if valid_ss58_format is not None and ss58_format != valid_ss58_format:
-        raise ValueError("Invalid SS58 format")
+    if valid_address_type and ss58_format[0] != valid_address_type:
+        raise ValueError("Invalid Address type")
 
     # Determine checksum length according to length of address string
-    if len(address_decoded) in [3, 4, 6, 10]:
+    if len(ss58_format) in [3, 4, 6, 10]:
         checksum_length = 1
-    elif len(address_decoded) in [5, 7, 11, 34 + ss58_format_length, 35 + ss58_format_length]:
+    elif len(ss58_format) in [5, 7, 11, 35]:
         checksum_length = 2
-    elif len(address_decoded) in [8, 12]:
+    elif len(ss58_format) in [8, 12]:
         checksum_length = 3
-    elif len(address_decoded) in [9, 13]:
+    elif len(ss58_format) in [9, 13]:
         checksum_length = 4
-    elif len(address_decoded) in [14]:
+    elif len(ss58_format) in [14]:
         checksum_length = 5
-    elif len(address_decoded) in [15]:
+    elif len(ss58_format) in [15]:
         checksum_length = 6
-    elif len(address_decoded) in [16]:
+    elif len(ss58_format) in [16]:
         checksum_length = 7
-    elif len(address_decoded) in [17]:
+    elif len(ss58_format) in [17]:
         checksum_length = 8
     else:
         raise ValueError("Invalid address length")
 
-    checksum = blake2b(checksum_prefix + address_decoded[0:-checksum_length]).digest()
+    checksum = blake2b(checksum_prefix + ss58_format[0:-checksum_length]).digest()
 
-    if checksum[0:checksum_length] != address_decoded[-checksum_length:]:
+    if checksum[0:checksum_length] != ss58_format[-checksum_length:]:
         raise ValueError("Invalid checksum")
 
-    return address_decoded[ss58_format_length:len(address_decoded)-checksum_length].hex()
+    return ss58_format[1:len(ss58_format)-checksum_length].hex()
 
 
-def ss58_encode(address: str, ss58_format: int = 42, address_type=None) -> str:
+def ss58_encode(address, address_type=42):
     """
     Encodes an account ID to an Substrate address according to provided address_type
 
     Parameters
     ----------
     address
-    ss58_format
-    address_type: (deprecated)
+    address_type
 
     Returns
     -------
@@ -114,19 +89,12 @@ def ss58_encode(address: str, ss58_format: int = 42, address_type=None) -> str:
     """
     checksum_prefix = b'SS58PRE'
 
-    if address_type is not None:
-        warnings.warn("Keyword 'address_type' will be replaced by 'ss58_format'", DeprecationWarning)
-        ss58_format = address_type
-
-    if ss58_format < 0 or ss58_format > 16383 or ss58_format in [46, 47]:
-        raise ValueError("Invalid value for ss58_format")
-
     if type(address) is bytes or type(address) is bytearray:
         address_bytes = address
     else:
         address_bytes = bytes.fromhex(address.replace('0x', ''))
 
-    if len(address_bytes) in [32, 33]:
+    if len(address_bytes) == 32:
         # Checksum size is 2 bytes for public key
         checksum_length = 2
     elif len(address_bytes) in [1, 2, 4, 8]:
@@ -135,108 +103,63 @@ def ss58_encode(address: str, ss58_format: int = 42, address_type=None) -> str:
     else:
         raise ValueError("Invalid length for address")
 
-    if ss58_format < 64:
-        ss58_format_bytes = bytes([ss58_format])
-    else:
-        ss58_format_bytes = bytes([
-            ((ss58_format & 0b0000_0000_1111_1100) >> 2) | 0b0100_0000,
-            (ss58_format >> 8) | ((ss58_format & 0b0000_0000_0000_0011) << 6)
-        ])
+    address_format = bytes([address_type]) + address_bytes
+    checksum = blake2b(checksum_prefix + address_format).digest()
 
-    input_bytes = ss58_format_bytes + address_bytes
-    checksum = blake2b(checksum_prefix + input_bytes).digest()
-
-    return base58.b58encode(input_bytes + checksum[:checksum_length]).decode()
+    return base58.b58encode(address_format + checksum[:checksum_length]).decode()
 
 
-def ss58_encode_account_index(account_index: int, ss58_format: int = 42, address_type=None) -> str:
+def ss58_encode_account_index(account_index, address_type=42):
     """
     Encodes an AccountIndex to an Substrate address according to provided address_type
 
     Parameters
     ----------
     account_index
-    ss58_format
-    address_type: (deprecated)
+    address_type
 
     Returns
     -------
 
     """
 
-    if address_type is not None:
-        warnings.warn("Keyword 'address_type' will be replaced by 'ss58_format'", DeprecationWarning)
-        ss58_format = address_type
-
-    if 0 <= account_index <= 2 ** 8 - 1:
-        account_idx_encoder = ScaleDecoder.get_decoder_class('u8')
-    elif 2 ** 8 <= account_index <= 2 ** 16 - 1:
-        account_idx_encoder = ScaleDecoder.get_decoder_class('u16')
-    elif 2 ** 16 <= account_index <= 2 ** 32 - 1:
-        account_idx_encoder = ScaleDecoder.get_decoder_class('u32')
-    elif 2 ** 32 <= account_index <= 2 ** 64 - 1:
-        account_idx_encoder = ScaleDecoder.get_decoder_class('u64')
+    if 0 <= account_index <= 2**8 - 1:
+        account_idx_encoder = U8()
+    elif 2**8 <= account_index <= 2**16 - 1:
+        account_idx_encoder = U16()
+    elif 2**16 <= account_index <= 2**32 - 1:
+        account_idx_encoder = U32()
+    elif 2**32 <= account_index <= 2**64 - 1:
+        account_idx_encoder = U64()
     else:
         raise ValueError("Value too large for an account index")
 
-    return ss58_encode(account_idx_encoder.encode(account_index).data, ss58_format)
+    return ss58_encode(account_idx_encoder.encode(account_index).data, address_type)
 
 
-def ss58_decode_account_index(address: str, valid_ss58_format: Optional[int] = None, valid_address_type=None) -> int:
+def ss58_decode_account_index(address, valid_address_type=42):
     """
     Decodes given SS58 encoded address to an AccountIndex
 
     Parameters
     ----------
     address
-    valid_ss58_format
     valid_address_type
 
     Returns
     -------
     Decoded int AccountIndex
     """
-
-    if valid_address_type is not None:
-        warnings.warn("Keyword 'valid_address_type' will be replaced by 'valid_ss58_format'", DeprecationWarning)
-        valid_ss58_format = valid_address_type
-
-    account_index_bytes = ss58_decode(address, valid_ss58_format)
+    account_index_bytes = ss58_decode(address, valid_address_type)
 
     if len(account_index_bytes) == 2:
-        return ScaleDecoder.get_decoder_class('u8', data=ScaleBytes('0x{}'.format(account_index_bytes))).decode()
+        return U8(ScaleBytes('0x{}'.format(account_index_bytes))).decode()
     if len(account_index_bytes) == 4:
-        return ScaleDecoder.get_decoder_class('u16', data=ScaleBytes('0x{}'.format(account_index_bytes))).decode()
+        return U16(ScaleBytes('0x{}'.format(account_index_bytes))).decode()
     if len(account_index_bytes) == 8:
-        return ScaleDecoder.get_decoder_class('u32', data=ScaleBytes('0x{}'.format(account_index_bytes))).decode()
+        return U32(ScaleBytes('0x{}'.format(account_index_bytes))).decode()
     if len(account_index_bytes) == 16:
-        return ScaleDecoder.get_decoder_class('u64', data=ScaleBytes('0x{}'.format(account_index_bytes))).decode()
+        return U64(ScaleBytes('0x{}'.format(account_index_bytes))).decode()
     else:
         raise ValueError("Invalid account index length")
 
-
-def is_valid_ss58_address(value: str, valid_ss58_format: Optional[int] = None) -> bool:
-    """
-    Checks if given value is a valid SS58 formatted address, optionally check if address is valid for specified
-    ss58_format
-
-    Parameters
-    ----------
-    value: value to checked
-    valid_ss58_format: if valid_ss58_format is provided the address must be valid for specified ss58_format (network) as well
-
-    Returns
-    -------
-    bool
-    """
-
-    # Return False in case a public key is provided
-    if value.startswith('0x'):
-        return False
-
-    try:
-        ss58_decode(value, valid_ss58_format=valid_ss58_format)
-    except ValueError:
-        return False
-
-    return True
