+++ bittensor/subtensor.py
@@ -15,6 +15,7 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
 # DEALINGS IN THE SOFTWARE.
 import argparse
+import asyncio
 import random
 import time
 
@@ -25,7 +26,7 @@ from termcolor import colored
 
 import bittensor
 import bittensor.utils.networking as net
-from bittensor.substrate.base import SubstrateInterface, Keypair
+from bittensor.substrate import SubstrateWSInterface, Keypair
 from bittensor.substrate.exceptions import SubstrateRequestException
 from bittensor.utils.neurons import Neuron, Neurons
 from bittensor.utils.balance import Balance
@@ -60,7 +61,7 @@ class Subtensor:
             wallet = bittensor.wallet.Wallet( self.config )
         self.wallet = wallet
 
-        self.substrate = SubstrateInterface(
+        self.substrate = SubstrateWSInterface(
             address_type = 42,
             type_registry_preset='substrate-node-template',
             type_registry=self.custom_type_registry,
@@ -79,7 +80,7 @@ class Subtensor:
     def add_args(parser: argparse.ArgumentParser):
         bittensor.wallet.Wallet.add_args( parser )
         try:
-            parser.add_argument('--subtensor.network', default='boltzmann', type=str, 
+            parser.add_argument('--subtensor.network', default='akira', type=str, 
                                 help='''The subtensor network flag. The likely choices are:
                                         -- akira (testing network)
                                         -- kusanagi (main network)
@@ -144,15 +145,35 @@ class Subtensor:
                 return None
             return random.choice( akira_available )
 
-    def is_connected(self):
+    def is_connected(self) -> bool:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_is_connected())
+
+    async def async_is_connected(self) -> bool:
         r""" Returns the connection state as a boolean.
         Raises:
             success (bool):
                 True is the websocket is connected to the chain endpoint.
         """
-        return self.substrate.is_connected()
+        return await self.substrate.async_is_connected()
+
+    def check_connection(self) -> bool:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_check_connection())
+
+    async def async_check_connection(self) -> bool:
+        if not await self.async_is_connected():
+            return await self.async_connect()
+        return True
 
     def connect(self, timeout: int = 10, failure = True ) -> bool:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_connect(timeout, failure))
+
+    async def async_connect(self, timeout: int = 10, failure = True ) -> bool:
         r""" Attempts to connect the substrate interface backend. 
         If the connection fails, attemps another endpoint until a timeout.
         Args:
@@ -168,20 +189,20 @@ class Subtensor:
         attempted_endpoints = []
         while True:
             def connection_error_message():
-                logger.error( '''   
-                                    Check that your internet connection is working and the chain endpoints are available: {}
-                                    The subtensor.network should likely be one of the following choices:
-                                        -- local -- (your locally running node)
-                                        -- akira (testnet)
-                                        -- kusanagi (mainnet)
-                                    Or you may set the endpoint manually using the --subtensor.chain_endpoint flag 
-                                    To connect run a local node (See: docs/running_a_validator.md)
-                              ''', attempted_endpoints)
+                print('''
+Check that your internet connection is working and the chain endpoints are available: {}
+The subtensor.network should likely be one of the following choices:
+    -- local -- (your locally running node)
+    -- akira (testnet)
+    -- kusanagi (mainnet)
+Or you may set the endpoint manually using the --subtensor.chain_endpoint flag 
+To run a local node (See: docs/running_a_validator.md) \n
+                              '''.format( attempted_endpoints) )
 
             # ---- Get next endpoint ----
             ws_chain_endpoint = self.endpoint_for_network( blacklist = attempted_endpoints )
             if ws_chain_endpoint == None:
-                logger.error('No more available endpoints for connection with subtensor.network: {} attempted: {}', self.config.subtensor.network, attempted_endpoints)
+                print(colored("No more available endpoints for connection with subtensor.network: {} attempted: {}".format(self.config.subtensor.network, attempted_endpoints), 'red'))
                 connection_error_message()
                 if failure:
                     raise RuntimeError('Unable to connect to network {}. Make sure your internet connection is stable and the network is properly set.'.format(self.config.subtensor.network))
@@ -190,14 +211,10 @@ class Subtensor:
             attempted_endpoints.append(ws_chain_endpoint)
 
             # --- Attempt connection ----
-            self.substrate.connect( ws_chain_endpoint )
-            time.sleep(1) # Give it a second to connect.
-
-            # ---- Success ----
-            if self.substrate.is_connected():
+            if await self.substrate.async_connect( ws_chain_endpoint, timeout = 3 ):
                 print(colored("Successfully connected to endpoint: {}".format(ws_chain_endpoint), 'green'))
                 return True
-
+            
             # ---- Timeout ----
             elif (time.time() - start_time) > timeout:
                 print(colored("Error while connecting to the chain endpoint {}".format(ws_chain_endpoint), 'red'))
@@ -207,19 +224,15 @@ class Subtensor:
                 else:
                     return False
 
-            # ---- Loop ----
-            else:
-                time.sleep(1)
+    def is_subscribed(self, ip: str, port: int, modality: int, coldkey: str) -> bool:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_is_subscribed( ip, port, modality, coldkey))
 
-    def _check_connection(self) -> bool:
-        if not self.is_connected():
-            return self.connect()
-        return True
-
-    def is_subscribed(self, ip: str, port: int, modality: int, coldkey: str):
-        uid = self.get_uid_for_pubkey(self.wallet.hotkey.public_key)
+    async def async_is_subscribed(self, ip: str, port: int, modality: int, coldkey: str) -> bool:
+        uid = await self.async_get_uid_for_pubkey(self.wallet.hotkey.public_key)
         if uid != None:
-            neuron = self.get_neuron_for_uid( uid )
+            neuron = await self.async_get_neuron_for_uid( uid )
             if neuron['ip'] == net.ip_to_int(ip) and neuron['port'] == port:
                 return True
             else:
@@ -228,12 +241,17 @@ class Subtensor:
             return False
 
     def subscribe(self, ip: str, port: int, modality: int, coldkeypub: str, wait_for_finalization = True) -> bool:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_subscribe( ip, port, modality, coldkeypub, wait_for_finalization))
+
+    async def async_subscribe(self, ip: str, port: int, modality: int, coldkeypub: str, wait_for_finalization = True) -> bool:
         r""" Subscribes the passed metadata to the substensor chain.
         """
-        if not self._check_connection():
+        if not await self.async_check_connection():
             return False
 
-        if self.is_subscribed( ip, port, modality, coldkeypub ):
+        if await self.async_is_subscribed( ip, port, modality, coldkeypub ):
             print(colored('Already subscribed with [ip: {}, port: {}, modality: {}, coldkey: {}]'.format(ip, port, modality, coldkeypub), 'green'))
             return True
 
@@ -245,39 +263,49 @@ class Subtensor:
             'modality': modality,
             'coldkey': coldkeypub,
         }
-        call = self.substrate.compose_call(
+        call = await self.substrate.compose_call(
             call_module='SubtensorModule',
             call_function='subscribe',
             call_params=params
         )
 
-        extrinsic = self.substrate.create_signed_extrinsic(call=call, keypair=self.wallet.hotkey)
+        extrinsic = await self.substrate.create_signed_extrinsic(call=call, keypair=self.wallet.hotkey)
         if wait_for_finalization:
             try:
-                print ('wait for finalization')
-                receipt = self.substrate.submit_extrinsic( extrinsic, wait_for_finalization = wait_for_finalization, timeout = bittensor.__blocktime__ * 5) # Waiting for block to finalized. 
-                print ('receipt', receipt)
-                if receipt.is_success:
-                    print(colored('Subscribed with [ip: {}, port: {}, modality: {}, coldkey: {}]'.format(ip, port, modality, coldkeypub), 'green'))
-                    return True
-                else:
-                    print(colored('Subscription failure with: error {}'.format(receipt.error_message), 'red'))
-                    return False
+                print ('waiting for finalization...')
+                receipt = await self.substrate.submit_extrinsic( extrinsic, wait_for_finalization = wait_for_finalization, timeout = bittensor.__blocktime__ * 5) # Waiting for block to finalized. 
+                try:
+                    if receipt.is_success:
+                        print(colored('Subscribed with [ip: {}, port: {}, modality: {}, coldkey: {}]'.format(ip, port, modality, coldkeypub), 'green'))
+                        return True
+                    else:
+                        print(colored('Subscription failure with: error {}'.format(receipt.error_message), 'red'))
+                        return False
+                except:
+                    False
 
             except SubstrateRequestException as e:
                 print(colored("Failed to send subscribe extrinsic with error: {}".format(e), 'red'))
         else:
             return True
 
-    def get_balance(self, address):
+    def get_balance(self, address) -> Balance:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_balance(address))
+
+    async def async_get_balance(self, address) -> Balance:
         r""" Returns the token balance for the passed ss58_address address
         Args:
             address (Substrate address format, default = 42):
                 ss58 chain address.
+        Return:
+            balance (bittensor.utils.balance.Balance):
+                account balance
         """
-        self._check_connection()
+        await self.async_check_connection()
         logger.debug("Getting balance for: {}", address)
-        result = self.substrate.get_runtime_state(
+        result = await self.substrate.get_runtime_state(
             module='System',
             storage_function='Account',
             params=[address],
@@ -292,6 +320,11 @@ class Subtensor:
         return Balance(balance)
 
     def add_stake(self, amount : Balance, hotkey_id: int):
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        loop.run_until_complete(self.async_add_stake(amount, hotkey_id))
+
+    async def async_add_stake(self, amount : Balance, hotkey_id: int):
         r""" Adds the specified amount of stake to passed hotkey uid.
         Args:
             amount (bittensor.utils.balance.Balance):
@@ -299,9 +332,9 @@ class Subtensor:
             hotkey_id (int):
                 uid of hotkey to stake into.
         """
-        self._check_connection()
+        await self.async_check_connection()
         logger.info("Adding stake of {} rao from coldkey {} to hotkey {}", amount.rao, self.wallet.coldkey.public_key, hotkey_id)
-        call = self.substrate.compose_call(
+        call = await self.substrate.compose_call(
             call_module='SubtensorModule',
             call_function='add_stake',
             call_params={
@@ -309,11 +342,16 @@ class Subtensor:
                 'ammount_staked': amount.rao
             }
         )
-        extrinsic = self.substrate.create_signed_extrinsic(call=call, keypair=self.wallet.coldkey)
-        result = self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=False)
+        extrinsic = await self.substrate.create_signed_extrinsic(call=call, keypair=self.wallet.coldkey)
+        result = await self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=False)
         return result
 
     def transfer(self, dest:str, amount: Balance):
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        loop.run_until_complete(self.async_transfer(dest, amount))
+
+    async def async_transfer(self, dest:str, amount: Balance):
         r""" Transfers funds from this wallet to the destination public key address
         Args:
             dest (str):
@@ -321,9 +359,9 @@ class Subtensor:
             amount (bittensor.utils.balance.Balance):
                 amount to stake as bittensor balance
         """
-        self._check_connection()
+        await self.async_check_connection()
         logger.debug("Requesting transfer of {}, from coldkey: {} to dest: {}", amount.rao, self.wallet.coldkey.public_key, dest)
-        call = self.substrate.compose_call(
+        call = await self.substrate.compose_call(
             call_module='Balances',
             call_function='transfer',
             call_params={
@@ -331,11 +369,16 @@ class Subtensor:
                 'value': amount.rao
             }
         )
-        extrinsic = self.substrate.create_signed_extrinsic(call=call, keypair = self.wallet.coldkey)
-        result = self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=False)
+        extrinsic = await self.substrate.create_signed_extrinsic(call=call, keypair = self.wallet.coldkey)
+        result = await self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=False)
         return result
 
     def unstake(self, amount : Balance, hotkey_id: int):
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        loop.run_until_complete(self.async_unstake(amount, hotkey_id))
+
+    async def async_unstake(self, amount : Balance, hotkey_id: int):
         r""" Removes stake into the wallet coldket from the specified hotkey uid.
         Args:
             amount (bittensor.utils.balance.Balance):
@@ -343,18 +386,23 @@ class Subtensor:
             hotkey_id (int):
                 uid of hotkey to unstake from.
         """
-        self._check_connection()
+        await self.async_check_connection()
         logger.debug("Requesting unstake of {} rao for hotkey: {} to coldkey: {}", amount.rao, hotkey_id, self.wallet.coldkey.public_key)
-        call = self.substrate.compose_call(
+        call = await self.substrate.compose_call(
             call_module='SubtensorModule',
             call_function='remove_stake',
             call_params={'ammount_unstaked': amount.rao, 'hotkey': hotkey_id}
         )
-        extrinsic = self.substrate.create_signed_extrinsic(call=call, keypair=self.wallet.coldkey)
-        result = self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=False)
+        extrinsic = await self.substrate.create_signed_extrinsic(call=call, keypair=self.wallet.coldkey)
+        result = await self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=False)
         return result
 
     def set_weights(self, destinations, values, wait_for_inclusion=False):
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        loop.run_until_complete(self.async_set_weights(destinations, values, wait_for_inclusion))
+
+    async def async_set_weights(self, destinations, values, wait_for_inclusion=False):
         r""" Sets the given weights and values on chain for wallet hotkey account.
         Args:
             destinations (List[int]):
@@ -362,91 +410,126 @@ class Subtensor:
             values (List[int]):
                 u32 max encoded floating point weights.
         """
-        self._check_connection()
-        call = self.substrate.compose_call(
+        await self.async_check_connection()
+        call = await self.substrate.compose_call(
             call_module='SubtensorModule',
             call_function='set_weights',
             call_params = {'dests': destinations, 'weights': values}
         )
-        extrinsic = self.substrate.create_signed_extrinsic(call=call, keypair = self.wallet.hotkey)
-        result = self.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=wait_for_inclusion)
+        extrinsic = await self.substrate.create_signed_extrinsic(call=call, keypair = self.wallet.hotkey)
+        result = await elf.substrate.submit_extrinsic(extrinsic, wait_for_inclusion=wait_for_inclusion)
         return result
 
     def get_current_block(self) -> int:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_current_block())
+
+    async def async_get_current_block(self) -> int:
         r""" Returns the block number from the chain.
         Returns:
             block_number (int):
                 Current chain blocknumber.
         """
-        self._check_connection()
-        return self.substrate.get_block_number(None)
+        await self.async_check_connection()
+        return await self.substrate.async_get_block_number(None)
 
     def get_active(self) -> List[Tuple[str, int]]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_active())
+
+    async def async_get_active(self) -> List[Tuple[str, int]]:
         r""" Returns a list of (public key, uid) pairs one for each active peer on chain.
         Returns:
             active (List[Tuple[str, int]]):
                 List of active peers.
         """
-        self._check_connection()
-        result =self.substrate.iterate_map(
+        await self.async_check_connection()
+        result = await self.substrate.iterate_map(
             module='SubtensorModule',
             storage_function='Active',
         )
         return result
 
     def get_stake(self) -> List[Tuple[int, int]]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_stake())
+
+    async def async_get_stake(self) -> List[Tuple[int, int]]:
         r""" Returns a list of (uid, stake) pairs one for each active peer on chain.
         Returns:
             stake (List[Tuple[int, int]]):
                 List of stake values.
         """
-        self._check_connection()
-        result =self.substrate.iterate_map(
+        await self.async_check_connection()
+        result = await self.substrate.iterate_map(
             module='SubtensorModule',
             storage_function='Stake',
         )
         return result
 
     def get_last_emit(self) -> List[Tuple[int, int]]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_last_emit())
+
+    async def async_get_last_emit(self) -> List[Tuple[int, int]]:
         r""" Returns a list of (uid, last emit) pairs for each active peer on chain.
         Returns:
             last_emit (List[Tuple[int, int]]):
                 List of last emit values.
         """
-        self._check_connection()
-        result = self.substrate.iterate_map(
+        await self.async_check_connection()
+        result = await self.substrate.iterate_map(
             module='SubtensorModule',
             storage_function='LastEmit'
         )
         return result
 
     def get_weight_vals(self) -> List[Tuple[int, List[int]]]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_weight_vals())
+
+    async def async_get_weight_vals(self) -> List[Tuple[int, List[int]]]:
         r""" Returns a list of (uid, weight vals) pairs for each active peer on chain.
         Returns:
             weight_vals (List[Tuple[int, List[int]]]):
                 List of weight val pairs.
         """
-        self._check_connection()
-        result = self.substrate.iterate_map(
+        await self.async_check_connection()
+        result = await self.substrate.iterate_map(
             module='SubtensorModule',
             storage_function='WeightVals'
         )
         return result
 
     def get_weight_uids(self) -> List[Tuple[int, int]]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_weight_uids())
+
+    async def async_get_weight_uids(self) -> List[Tuple[int, int]]:
         r""" Returns a list of (uid, weight uids) pairs for each active peer on chain.
         Returns:
             weight_uids (List[Tuple[int, int]]):
                 List of weight uid pairs
         """
-        self._check_connection()
-        result = self.substrate.iterate_map(
+        await self.async_check_connection()
+        result = await self.substrate.iterate_map(
             module='SubtensorModule',
             storage_function='WeightUids'
         )
         return result
 
-    def neurons(self, uid=None, decorator=False):
+    def neurons(self, uid=None, decorator=False) -> List[Tuple[int, dict]]: 
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_neurons( uid, decorator ))
+
+    async def async_neurons(self, uid=None, decorator=False) -> List[Tuple[int, dict]]:
         r""" Returns a list of neuron from the chain. 
             Returns neuron objects if decorator is set to true.
         Args:
@@ -455,13 +538,13 @@ class Subtensor:
             decorator(bool):
                 specifiying if the returned data should be as neuron objects.
         Returns:
-            neuron (List[Tuple[int, Dict]]):
+            neuron (List[Tuple[int, dict]]):
                 List of neuron objects.
         """
-        self._check_connection()
+        await self.async_check_connection()
         # Todo (parall4x, 23-12-2020) Get rid of this decorator flag. This should be refactored into something that returns Neuron objects only
         if uid:
-            result = self.substrate.get_runtime_state(
+            result = await self.substrate.get_runtime_state(
                 module='SubtensorModule',
                 storage_function='Neurons',
                 params=[uid]
@@ -469,13 +552,18 @@ class Subtensor:
             return Neurons.from_list(result['result']) if decorator else result['result']
 
         else:
-            neurons = self.substrate.iterate_map(
+            neurons = await self.substrate.iterate_map(
                 module='SubtensorModule',
                 storage_function='Neurons'
             )
             return Neurons.from_list(neurons) if decorator else neurons
 
-    def get_uid_for_pubkey(self, pubkey = str) -> int:
+    def get_uid_for_pubkey(self, pubkey = str) -> int: 
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_uid_for_pubkey( pubkey ))
+
+    async def async_get_uid_for_pubkey(self, pubkey = str) -> int:
         r""" Returns the uid of the peer given passed public key string.
         Args:
             pubkey (str):
@@ -484,8 +572,8 @@ class Subtensor:
             uid (int):
                 uid of peer with hotkey equal to passed public key.
         """
-        self._check_connection()
-        result = self.substrate.get_runtime_state(
+        await self.async_check_connection()
+        result = await self.substrate.get_runtime_state(
             module='SubtensorModule',
             storage_function='Active',
             params=[pubkey]
@@ -494,7 +582,12 @@ class Subtensor:
             return None
         return int(result['result'])
 
-    def get_neuron_for_uid(self, uid:int):
+    def get_neuron_for_uid(self, uid) -> dict:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_neuron_for_uid( uid ))
+
+    async def async_get_neuron_for_uid(self, uid:int) -> dict:
         r""" Returns the neuron metadata of the peer with the passed uid.
         Args:
             uid (int):
@@ -503,16 +596,20 @@ class Subtensor:
             metadata (Dict):
                 Dict in list form containing metadata of associated uid.
         """
-        self._check_connection()
-        result = self.substrate.get_runtime_state(
+        await self.async_check_connection()
+        result = await self.substrate.get_runtime_state(
                 module='SubtensorModule',
                 storage_function='Neurons',
                 params=[uid]
         )
         return result['result']
 
-
     def get_stake_for_uid(self, uid) -> Balance:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_stake_for_uid( uid ))
+
+    async def async_get_stake_for_uid(self, uid) -> Balance:
         r""" Returns the staked token amount of the peer with the passed uid.
         Args:
             uid (int):
@@ -521,8 +618,8 @@ class Subtensor:
             stake (int):
                 Amount of staked token.
         """
-        self._check_connection()
-        stake = self.substrate.get_runtime_state(
+        await self.async_check_connection()
+        stake = await self.substrate.get_runtime_state(
             module='SubtensorModule',
             storage_function='Stake',
             params = [uid]
@@ -532,7 +629,12 @@ class Subtensor:
             return Balance(0)
         return Balance(result)
 
-    def weight_uids_for_uid(self, uid):
+    def weight_uids_for_uid(self, uid) -> List[int]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_weight_uids_for_uid( uid ))
+
+    async def async_weight_uids_for_uid(self, uid) -> List[int]:
         r""" Returns the weight uids of the peer with the passed uid.
         Args:
             uid (int):
@@ -541,15 +643,20 @@ class Subtensor:
             weight_uids (List[int]):
                 Weight uids for passed uid.
         """
-        self._check_connection()
-        result = self.substrate.get_runtime_state(
+        await self.async_check_connection()
+        result = await self.substrate.get_runtime_state(
             module='SubtensorModule',
             storage_function='WeightUids',
             params = [uid]
         )
         return result['result']
 
-    def weight_vals_for_uid(self, uid):
+    def weight_vals_for_uid(self, uid) -> List[int]:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_weight_vals_for_uid( uid ))
+
+    async def async_weight_vals_for_uid(self, uid) -> List[int]:
         r""" Returns the weight vals of the peer with the passed uid.
         Args:
             uid (int):
@@ -558,15 +665,20 @@ class Subtensor:
             weight_vals (List[int]):
                 Weight vals for passed uid.
         """
-        self._check_connection()
-        result = self.substrate.get_runtime_state(
+        await self.async_check_connection()
+        result = await self.substrate.get_runtime_state(
             module='SubtensorModule',
             storage_function='WeightVals',
             params = [uid]
         )
         return result['result']
 
-    def get_last_emit_data_for_uid(self, uid):
+    def get_last_emit_data_for_uid(self, uid) -> int:
+        loop = asyncio.get_event_loop()
+        loop.set_debug(enabled=True)
+        return loop.run_until_complete(self.async_get_last_emit_data_for_uid( uid ))
+
+    async def async_get_last_emit_data_for_uid(self, uid) -> int:
         r""" Returns the last emit of the peer with the passed uid.
         Args:
             uid (int):
@@ -575,39 +687,10 @@ class Subtensor:
             last_emit (int):
                 Last emit block numebr
         """
-        self._check_connection()
-        result = self.substrate.get_runtime_state(
+        await self.async_check_connection()
+        result = await self.substrate.get_runtime_state(
             module='SubtensorModule',
             storage_function='LastEmit',
             params = [uid]
         )
         return result['result']
-
-
-    def get_full_state(self) -> Tuple[List[int], List[Tuple[int, dict]], List[int], List[Tuple[int, List[int]]],  List[Tuple[int, List[int]]], List[Tuple[int, str]]]:
-        self._check_connection()
-        last_emit = self.substrate.iterate_map(
-            module='SubtensorModule',
-            storage_function='LastEmit'
-        )
-        neurons = self.substrate.iterate_map(
-            module='SubtensorModule',
-            storage_function='Neurons'
-        )
-        stake = self.substrate.iterate_map(
-            module='SubtensorModule',
-            storage_function='Stake'
-        )
-        weight_vals = self.substrate.iterate_map(
-            module='SubtensorModule',
-            storage_function='WeightVals'
-        )
-        weight_uids = self.substrate.iterate_map(
-            module='SubtensorModule',
-            storage_function='WeightUids'
-        )
-        active = self.substrate.iterate_map(
-            module='SubtensorModule',
-            storage_function='Active'
-        )
-        return last_emit, neurons, stake, weight_vals, weight_uids, active
