+++ bittensor/utils/cli_utils.py
@@ -8,7 +8,7 @@ import getpass
 
 from loguru import logger
 
-from bittensor.substrate.base import Keypair
+from bittensor.substrate import Keypair
 from bittensor.crypto.keyfiles import load_keypair_from_data, KeyFileError
 from termcolor import colored
 from bittensor.crypto import encrypt, is_encrypted, decrypt_data, KeyError
