+++ bittensor/wallet.py
@@ -74,13 +74,13 @@ class Wallet():
         pass
 
     @property
-    def hotkey(self) -> bittensor.substrate.base.Keypair:
+    def hotkey(self) -> bittensor.substrate.Keypair:
         if self._hotkey == None:
             self._load_hotkey()
         return self._hotkey
 
     @property
-    def coldkey(self) -> bittensor.substrate.base.Keypair:
+    def coldkey(self) -> bittensor.substrate.Keypair:
         if self._coldkey == None:
             self._load_coldkey( )
         return self._coldkey
@@ -208,8 +208,8 @@ class Wallet():
         return st.st_mode & stat.S_IROTH
 
     @staticmethod
-    def __create_keypair() -> bittensor.substrate.base.Keypair:
-        return bittensor.substrate.base.Keypair.create_from_mnemonic(bittensor.substrate.base.Keypair.generate_mnemonic())
+    def __create_keypair() -> bittensor.substrate.Keypair:
+        return bittensor.substrate.Keypair.create_from_mnemonic(bittensor.substrate.Keypair.generate_mnemonic())
 
     @staticmethod
     def __save_keypair(keypair : bittensor.substrate.Keypair, path : str):
