+++ tests/unit_tests/subtensor/interface/test_helper_functions.py
@@ -21,7 +21,7 @@ from unittest.mock import MagicMock
 from scalecodec import ScaleBytes, Bytes
 from scalecodec.metadata import MetadataDecoder
 
-from bittensor.substrate.base import SubstrateInterface
+from bittensor.substrate import SubstrateWSInterface
 from .fixtures import metadata_v12_hex
 
 
@@ -30,7 +30,7 @@ class TestHelperFunctions(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
 
-        cls.substrate = SubstrateInterface(host="dummy", port=666, address_type=42, type_registry_preset='kusama')
+        cls.substrate = SubstrateWSInterface(address_type=42, type_registry_preset='kusama')
         metadata_decoder = MetadataDecoder(ScaleBytes(metadata_v12_hex))
         metadata_decoder.decode()
         cls.substrate.get_block_metadata = MagicMock(return_value=metadata_decoder)
